requirements.txt
google-cloud-bigquery==1.8.1